emacs: iterate on daily template and roam keybinds
Daily template now includes a "highlights" section to track anything
of value. The intent is to lift these into the weekly document to help
summarize what was accomplished in a given week.

Experimenting with using the default keymap for org-roam, to capture
daily notes, rather than manually adding them to the daily file. Not
sure if this will stick around.